refactor: break out PercentageBarGroup to its own file
