+++ apps/src/code-studio/pd/workshop_dashboard/workshops/workshop.module.scss
@@ -400,39 +400,6 @@ div.card {
           background-color: var(--background-error-primary);
         }
       }
-
-      .barRow {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        gap: 0.5rem;
-
-        .barLabel {
-          min-width: 5rem;
-          flex-shrink: 0;
-          text-align: left;
-          color: var(--text-neutral-tertiary);
-        }
-
-        .barContainer {
-          position: relative;
-          flex: 1;
-          flex-shrink: 0;
-          height: 0.625rem;
-          border-radius: 6.25rem;
-          background: var(--background-neutral-tertiary);
-          overflow: hidden;
-
-          .indicator {
-            background: var(--background-brand-teal-primary);
-            position: absolute;
-            left: 0;
-            top: 0;
-            bottom: 0;
-            transition: width 0.6s ease-out;
-          }
-        }
-      }
     }
 
     .scoreCardFooter {
